Sort selectors before grouping them
This is actually how group_by works. The iterator needs to first be sorted. To do this I've added
a  method that returns the string representation of the SelectorWithOrginRule's selctor.
This is what we want to sort and group things by so seems applicable. It's a bit gross converting
to a string but it avoids having to do some weird implementation of Ord for prometheus_parser::Selector.