Handle errors parsing Prometheus expressions
Previously if we failed to parse an expression there would be a log at
error level but the exit code would still be zero - not what we want!
Now the errors are logged and failure is indicated so that once we've
finished processing the other files we can bubble the error up.