Allow specifying an evaluation interval for the rules
If there are lot of these rules then it can be expensive to evaluate them
regularly. For example, we evaluate rules every 15s by default. You also
probably don't need to be alerted in as timely manner for absent alerts.
This allows you to specify a custom interval at which to evaluate the
generated absent rules.